Add global postal fee conditions support
These postal fee conditions should be used for postal
fees, which should be somehow limited globally and
always, opposing to conditions which currently
needed to be selected in the country postal fee form
in the CRM admin.

Globally registered condition will always be
evaluated.

Interface provides "affectedPostalFees" and "adminMessages"
methods so we can display information about this limitation
in the CRM admin form.

remp/crm#3516